get con token access a url user_profile/backend, devuelve usuario logeado
