+++ frontend/src/app/page.tsx
@@ -1,7 +1,4 @@
-import Image from "next/image";
-import Footer from "./component/Footer";
-import Login from "./login/page";
-import Header from "./component/header";
+
 import AuthComponent from "./register_user/page";
